Delay documentation deployment
This should hopefully fix any possible issues where the NPM package
may be released after the documentation is deployed, which could cause
a cause miss by jsDelivr and cause the documentation to cache without
pulling in the Select2 dependency.

It hasn't happened yet, but I'm concerned it could happen and I don't
want to put in the effort to make a more sophisticated way of
handling this issue.